requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# OpenAI ve LangChain (tiktoken olmadan)
langchain==0.0.100
openai==0.28.1

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Environment Variables
python-dotenv==1.0.0
